% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{callAggregationMethod}
\alias{callAggregationMethod}
\title{Send a call to function in the server and get the result}
\usage{
callAggregationMethod(datasources, function_call)
}
\arguments{
\item{datasources}{a list of \code{\link{DSConnection-class}}
objects obtained after login.}

\item{function_call}{a character string that represents the call to a function with its arguments.}
}
\value{
\code{callAggregationMethod} returns the result from a call to a function in the server
}
\description{
Send a call to function in the server and get the result
}
