% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boundedCovarianceDP.R
\name{ds.boundedCovarianceDP}
\alias{ds.boundedCovarianceDP}
\title{Differentially private covariance}
\usage{
ds.boundedCovarianceDP(x, y, epsilon, x_min, x_max, y_min, y_max, datasources)
}
\arguments{
\item{x}{First input to the covariance}

\item{y}{Second input to the covariance}

\item{epsilon}{Privacy budget}

\item{x_min}{Lower bound for x}

\item{x_max}{Upper bound for x}

\item{y_min}{Lower bound for y}

\item{y_max}{Upper bound for y}

\item{datasources}{a list of \code{\link{DSConnection-class}}
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
\code{ds.boundedCovarianceDP} returns a differentially private covariance
}
\description{
Differentially private covariance
}
