% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boundedMeanDP.R
\name{ds.boundedMeanDP}
\alias{ds.boundedMeanDP}
\title{Title}
\usage{
ds.boundedMeanDP(datasources, input_data, epsilon, lower_bound, upper_bound)
}
\arguments{
\item{datasources}{}
}
\description{
Title
}
