% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boundedVarianceDP.R
\name{ds.boundedVarianceDP}
\alias{ds.boundedVarianceDP}
\title{Differentially private variance}
\usage{
ds.boundedVarianceDP(
  input_data,
  epsilon,
  lower_bound,
  upper_bound,
  datasources = NULL,
  type = "combine"
)
}
\arguments{
\item{input_data}{the input vector}

\item{epsilon}{privacy budget}

\item{lower_bound}{lower bound for input values}

\item{upper_bound}{upper bound for input values}

\item{datasources}{a list of \code{\link{DSConnection-class}}
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}

\item{type}{a character string that represents the type of analysis to carry out.
This can be set as \code{'combine'}, \code{'split'} or \code{'both'}.}
}
\value{
\code{ds.boundedVarianceDP} returns a differentially private covariance
}
\description{
Differentially private variance
}
