% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.studentTestDP.R
\name{ds.studentTestDP}
\alias{ds.studentTestDP}
\title{Differentially private student-test}
\usage{
ds.studentTestDP(
  datasources,
  x,
  y,
  epsilon,
  x_min,
  x_max,
  y_min,
  y_max,
  type = "combine"
)
}
\arguments{
\item{datasources}{a list of \code{\link{DSConnection-class}}
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}

\item{x}{First group to compare with the student-test}

\item{y}{Second group to compare with the student-test}

\item{epsilon}{Privacy budget}

\item{x_min}{Lower bound for x}

\item{x_max}{Upper bound for x}

\item{y_min}{Lower bound for y}

\item{y_max}{Upper bound for y}
}
\value{
\code{ds.studentTestDP} returns a differentially private student-test
}
\description{
Differentially private student-test
}
