% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boundedStandardDeviationDP.R
\name{ds.boundedStandardDeviationDP}
\alias{ds.boundedStandardDeviationDP}
\title{Title}
\usage{
ds.boundedStandardDeviationDP(
  datasources,
  input_data,
  epsilon,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{datasources}{}
}
\description{
Title
}
