% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boundedSumDP.R
\name{ds.boundedSumDP}
\alias{ds.boundedSumDP}
\title{Title}
\usage{
ds.boundedSumDP(
  datasources,
  input_data,
  epsilon,
  delta,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{datasources}{}
}
\description{
Title
}
